Updated project.py and my.py
Updated project.py to check the Involvement table to see if user is assigned to the project and to use POST request for updating task statuses. Updated my.py to have dashboard function as myProjects did previously. Also updated to only show the projects that the current logged in user is assigned to.